fix: set initial window size to large resolution.
This only takes effect until the window receives a real size from the
window manager/compositor.  With Hyprland (and probably all tiling
compositors) that occurs as soon as the window is tiled; probably with
a floating setup the window will actually display at these dimensions
until resized, so we might prefer e.g. 1024.786.

On my setup setting a large resolution here gives the window
sufficient detail to avoid blurring when it first opens.